+++ src/main/scala/com/github/plume/oss/Jimple2Cpg.scala
@@ -141,6 +141,11 @@ class Jimple2Cpg {
           .map(_.toString)
           .toSet[String]
 
+        driver match {
+          case x: OverflowDbDriver => x.removeExpiredPathsFromCache(unchangedTypes)
+          case _                   =>
+        }
+
         new PlumeMetaDataPass(cpg, language, Some(metaDataKeyPool), unchangedTypes)
           .createAndApply(driver)
 
@@ -162,10 +167,6 @@ class Jimple2Cpg {
         controlFlowPasses(cpg).foreach(_.createAndApply(driver))
         new PlumeReachingDefPass(cpg, unchangedTypes = unchangedTypes).createAndApply(driver)
         new PlumeHashPass(cpg).createAndApply(driver)
-        driver match {
-          case x: OverflowDbDriver => x.removeExpiredPathsFromCache(unchangedTypes)
-          case _                   =>
-        }
 
         driver.buildInterproceduralEdges()
         cpg